test: split server-components tests out of basic into dedicated fixture
