+++ test/fixtures/server-components/app/components/ComponentWithRef.vue
@@ -1,0 +1,16 @@
+<script setup lang="ts">
+const list = ['a', 'b', 'c']
+const templateRef = ref<HTMLElement[] | null>(null)
+</script>
+
+<template>
+  <div>
+    <div
+      v-for="item in list"
+      ref="templateRef"
+      :key="item"
+    >
+      {{ item }}
+    </div>
+  </div>
+</template>
