+++ test/fixtures/server-components/app/components/CounterWithNuxtClient.vue
