+++ test/fixtures/server-components/app/components/ServerInner.server.vue
@@ -1,0 +1,8 @@
+<template>
+  <div class="server-inner-counter">
+    <Counter
+      nuxt-client
+      :multiplier="1"
+    />
+  </div>
+</template>
