+++ test/fixtures/server-components/app/components/ServerOuter.server.vue
@@ -1,0 +1,5 @@
+<template>
+  <div class="server-outer">
+    <ServerInner />
+  </div>
+</template>
