+++ test/fixtures/server-components/app/components/ServerOuterUniversalNested.server.vue
@@ -1,0 +1,5 @@
+<template>
+  <div class="server-outer-universal-nested">
+    <UniversalWithNuxtClient />
+  </div>
+</template>
