+++ test/fixtures/server-components/app/components/ServerWithClient.server.vue
