+++ test/fixtures/server-components/app/components/ServerWithNestedClient.server.vue
