+++ test/fixtures/server-components/app/components/SlotWrapper.vue
@@ -1,0 +1,11 @@
+<template>
+  <div class="slot-wrapper">
+    <SlottedChild v-bind="$attrs">
+      <slot />
+    </SlottedChild>
+  </div>
+</template>
+
+<script setup lang="ts">
+defineOptions({ inheritAttrs: false })
+</script>
