+++ test/fixtures/server-components/app/components/UniversalWithNuxtClient.vue
@@ -1,0 +1,8 @@
+<template>
+  <div class="universal-counter">
+    <Counter
+      nuxt-client
+      :multiplier="1"
+    />
+  </div>
+</template>
