+++ test/fixtures/server-components/app/components/islands/LongAsyncComponent.vue
