+++ test/fixtures/server-components/app/components/islands/PureComponent.vue
