+++ test/fixtures/server-components/app/components/islands/RouteComponent.vue
