+++ test/fixtures/server-components/app/error.vue
@@ -1,0 +1,14 @@
+<template>
+  <div>
+    <h1>{{ error?.message }}</h1>
+    This is the error page 😱
+  </div>
+</template>
+
+<script setup lang="ts">
+import type { NuxtError } from '#app'
+
+defineProps({
+  error: Object as () => NuxtError,
+})
+</script>
