+++ test/fixtures/server-components/app/pages/islands.vue
