+++ test/fixtures/server-components/app/pages/nested-nuxt-client.vue
@@ -1,0 +1,6 @@
+<template>
+  <div>
+    <ServerOuter />
+    <ServerOuterUniversalNested />
+  </div>
+</template>
