+++ test/fixtures/server-components/app/pages/prefetch/server-components.vue
@@ -1,0 +1,7 @@
+<template>
+  <div>
+    Testing prefetching server components:
+
+    <AsyncServerComponent :count="34" />
+  </div>
+</template>
