+++ test/fixtures/server-components/app/pages/selective-client-slots.vue
@@ -1,0 +1,7 @@
+<template>
+  <SlotWrapper>
+    <button id="slotted">
+      click
+    </button>
+  </SlotWrapper>
+</template>
