+++ test/fixtures/server-components/app/pages/server-components/lazy/end.vue
