+++ test/fixtures/server-components/app/pages/server-components/lazy/start.vue
