+++ test/fixtures/server-components/app/pages/server-components/lost-page.server.vue
