+++ test/fixtures/server-components/app/pages/server-page.server.vue
