+++ test/fixtures/server-components/app/plugins/title-template.ts
@@ -1,0 +1,10 @@
+export default defineNuxtPlugin(() => {
+  useHead({
+    titleTemplate: '%s - Fixture',
+    style: [{
+      innerHTML: () => 'body { color: red }',
+      tagPriority: -2,
+      id: 'plugin-style',
+    }],
+  })
+})
