+++ test/fixtures/server-components/nuxt.config.ts
@@ -1,0 +1,17 @@
+import { withMatrix } from '../../matrix'
+
+export default withMatrix({
+  experimental: {
+    componentIslands: {
+      selectiveClient: 'deep',
+    },
+    runtimeBaseURL: true,
+  },
+  nitro: {
+    prerender: {
+      routes: [
+        '/prefetch/server-components',
+      ],
+    },
+  },
+})
