+++ test/fixtures/server-components/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "fixture-server-components",
+  "private": true,
+  "type": "module",
+  "dependencies": {
+    "nuxt": "workspace:*"
+  },
+  "engines": {
+    "node": "^20.19.0 || >=22.12.0"
+  }
+}
