+++ test/fixtures/server-components/server/api/very-long-request.ts
@@ -1,0 +1,9 @@
+import { defineEventHandler } from 'h3'
+
+export default defineEventHandler(async () => {
+  await timeout(20)
+  return 'that was very long ...'
+})
+function timeout (ms: number) {
+  return new Promise(resolve => setTimeout(resolve, ms))
+}
